refactor: move manual summarization helpers
